% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{coalesce_identifiers}
\alias{coalesce_identifiers}
\title{Find identifiers from a list of files using patterns and coalesce into single identifier}
\usage{
coalesce_identifiers(file, file_pattern)
}
\arguments{
\item{file_pattern}{}
}
\description{
Find identifiers from a list of files using patterns and coalesce into single identifier
}
