% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_identifier}
\alias{extract_identifier}
\title{Extract identifiers from a list of files using regexes}
\usage{
extract_identifier(file_list = list_files(), file_pattern = ".*")
}
\arguments{
\item{file_list}{A list of files from \code{\link[fs:dir_ls]{fs::dir_info()}}.}

\item{file_pattern}{A list of patterns to match for files and extract capture groups.}
}
\value{
A \code{file_list} with a new column, \code{identifier}, populated using the regexes of the
parameter \code{file_pattern}.
}
\description{
Extract identifiers from a list of files using regexes
}
\examples{
if (FALSE) {
  extract_identifier(file_list=list_files("."), file_pattern="^([A-Z]).*")
}
}
