% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileFinder-package.R
\docType{package}
\name{FileFinder-package}
\alias{FileFinder}
\alias{FileFinder-package}
\title{FileFinder: simply finding files}
\description{
FileFinder provides a set of basic functions useful when finding files of
specific characteristics like most recent modification or biggest / smallest
}
\author{
\strong{Maintainer}: Konrad Zdeb \email{konrad.zdeb@me.com}

}
\keyword{internal}
