% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_files}
\alias{find_files}
\title{Find files from a list of files using filters}
\usage{
find_files(file_list = list_files(), dirs = ".", file_pattern = ".*")
}
\arguments{
\item{file_list}{A list of files from \code{\link[fs:dir_ls]{fs::dir_info()}}.}

\item{dirs}{A list of directories to filter on (can be absolute, use regex)}

\item{file_pattern}{A list of patterns to match for files}
}
\value{
A \code{file_list} filtered to the files matching the criteria.
}
\description{
Find files from a list of files using filters
}
\examples{
if (FALSE) {
  find_files(file_list=list_files("."), dirs = ".", file_pattern=".*")
}
}
