% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{find_identifier}
\alias{find_identifier}
\title{Find identifier from a list of files using a single pattern}
\usage{
find_identifier(file, file_pattern)
}
\arguments{
\item{file_pattern}{}
}
\description{
Find identifier from a list of files using a single pattern
}
