% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_identifiers}
\alias{find_identifiers}
\title{Find identifiers from a list of files using patterns}
\usage{
find_identifiers(file, file_pattern)
}
\arguments{
\item{file_pattern}{}
}
\description{
Find identifiers from a list of files using patterns
}
