% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_identifier.R
\name{infer_identifier}
\alias{infer_identifier}
\title{Find files from a list of files using filters}
\usage{
infer_identifier(file_list = list_files(), file_pattern = ".*")
}
\arguments{
\item{file_list}{A list of files from \code{\link[fs:dir_ls]{fs::dir_info()}}.}

\item{file_pattern}{A list of patterns to match for files}

\item{dirs}{A list of directories to filter on (can be absolute, use regex)}
}
\value{
A \code{file_list} filtered to the files matching the criteria.
}
\description{
Find files from a list of files using filters
}
\examples{
if (FALSE) {
  find_files(file_list=list_files("."), dirs = ".", file_pattern=".*")
}
}
