% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files in one or more directories}
\usage{
list_files(path = ".")
}
\arguments{
\item{path}{A character vector of one or more paths. See \code{\link[fs:dir_ls]{fs::dir_info()}} for more details.}
}
\value{
A tibble of file information
}
\description{
Performs a \code{\link[fs:dir_ls]{fs::dir_info()}} function call on the specified paths
with recursion.
}
\examples{
if (FALSE) {
   list_files(".")
}
}
