% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_file_path.R
\name{split_file_path}
\alias{split_file_path}
\title{Split filename into file/path components}
\usage{
split_file_path(file_list = list_files())
}
\arguments{
\item{file_list}{A tibble of file information from \code{\link[fs:dir_ls]{fs::dir_info()}}.}
}
\value{
A \code{file_list} with two additional columns: \code{dir} and \code{file}.
}
\description{
Split filename into file/path components
}
