% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_file.R
\name{most_recent_file}
\alias{most_recent_file}
\title{Most Recent File}
\usage{
most_recent_file(path, n = 1, ...)
}
\arguments{
\item{path}{Valid folder path}

\item{n}{Number of files to return defulats to 1; as in \code{\link[base]{head}}.}

\item{...}{As in \code{\link[base]{list.files}}.}
}
\value{
A vector with file paths of length \code{n}.
}
\description{
Most Recent File
}
\examples{
most_recent_file(tempdir())
}
